+++ src/irvine/oeis/a000/A000686.java
@@ -21,7 +21,7 @@ public class A000686 extends A000685 {
     mC.add(super.next().multiply(3));
     Z s = Z.ZERO;
     for (int j = 0; j <= mN; ++j) {
-      s = s.add(Binomial.binomial(mN, j).shiftLeft(j * (mN - j)).multiply(mC.get(j)));
+      s = s.add(Binomial.binomial(mN, j).shiftLeft((long) j * (mN - j)).multiply(mC.get(j)));
     }
     return s.shiftRight(2);
   }
