+++ src/irvine/oeis/a001/A001308.java
@@ -15,8 +15,8 @@ public class A001308 extends Sequence1 {
   public Z next() {
     Z p = Z.ONE.shiftLeft(++mN).subtract(1);
     for (int k = 1; k < mN; ++k) {
-      p = p.multiply(Z.ONE.shiftLeft(2 * k).subtract(1));
+      p = p.multiply(Z.ONE.shiftLeft(2L * k).subtract(1));
     }
-    return p.shiftLeft(mN * (mN - 1));
+    return p.shiftLeft((long) mN * (mN - 1));
   }
 }
