+++ src/irvine/oeis/a006/A006308.java
@@ -112,7 +112,7 @@ public class A006308 extends AbstractSequence {
   @Override
   public Z next() {
     mP = (int) mPrime.nextPrime(mP);
-    final long p2 = mP * mP;
+    final long p2 = (long) mP * mP;
     final Z g = ZUtils.leastPrimitiveRoot(Z.valueOf(p2));
     final int q = g.modPow(Z.valueOf(mP), Z.valueOf(p2)).intValueExact();
     final Polynomial<CR> f = f(mP, q);
